feat: initial commit

+++ demo/src/main/java/com/example/minionlinemarket/Repository/LineItemRepo.java
@@ -5,8 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LineItemRepo extends JpaRepository<LineItem, Integer> {
-
-
+public interface LineItemRepo extends JpaRepository<LineItem, Long> {
 
 }
